Added button disabling/enabling.
Whenever you enter a new area it checks which directions you can go and disables and enables buttons accordingly.  I also added some more comments to clear things up.  With the addition of button disabling I decided to replace the function that checked if the user can go in requested area with a line that checks if the button is enabled in that direction.